Add first version
This file was built by taking all the SDC name-CUSIP pairs without a match in the main repo's mapping and compare them to all Compustat-GVKEY pairs.  The match is done on SDC names-Compustat names. Note that this is a time-invariant mapping (so there are multiple Compustat IDs to different SDC ids).  From Daniel Yang (OSU).